requirements.txt
numpy==1.22
tensorflow_gpu==2.7.0
keras==2.7.0
matplotlib==3.3.4
tensorflow-addons==0.15.0
scikit-learn==0.22.2
opencv-python==4.10.0.84
vit_keras==0.1.2
statsmodels==0.14.4